+++ apps/vc-api/src/vc-api/exchanges/entities/transaction.entity.ts
@@ -27,6 +27,7 @@ import { PresentationReviewEntity } from './presentation-review.entity';
 import { VpRequestEntity } from './vp-request.entity';
 import { CallbackConfiguration } from '../types/callback-configuration';
 import { PresentationSubmissionEntity } from './presentation-submission.entity';
+import { VpRequestSubmissionVerifier } from '../vp-request-submission-verifier';
 
 /**
  * A TypeOrm entity representing an exchange transaction
@@ -98,75 +99,17 @@ export class TransactionEntity {
   @Column('simple-json')
   callback: CallbackConfiguration[];
 
-  private verifyVpRequestTypeDidAuth(presentation: VerifiablePresentation): string[] {
-    // https://w3c-ccg.github.io/vp-request-spec/#did-authentication-request
-    const errors: string[] = [];
-
-    if (!presentation.holder) {
-      errors.push('Presentation holder is required for didAuth query');
-    }
-
-    return errors;
-  }
-
-  private verifyVpRequestTypePresentationDefinition(
-    presentation: VerifiablePresentation,
-    credentialQuery: Array<{ presentationDefinition: IPresentationDefinition }>
-  ): string[] {
-    // https://identity.foundation/presentation-exchange/#presentation-definition
-    const errors: string[] = [];
-    const pex: PEX = new PEX();
-
-    credentialQuery.forEach(({ presentationDefinition }, index) => {
-      const { errors: partialErrors } = pex.evaluatePresentation(
-        presentationDefinition,
-        presentation as IPresentation
-      );
-
-      errors.push(
-        ...partialErrors.map(
-          (error) =>
-            `Presentation definition (${index + 1}) validation failed, reason: ${error.message || 'Unknown'}`
-        )
-      );
-    });
-
-    return errors;
-  }
-
-  private validatePresentation(presentation: VerifiablePresentation): string[] {
-    const commonErrors = [];
-    // Common checking
-    if (presentation.proof.challenge !== this.vpRequest.challenge) {
-      commonErrors.push('Challenge does not match');
-    }
-
-    // Type specific checking
-    const partialErrors = this.vpRequest.query.flatMap((vpQuery) => {
-      switch (vpQuery.type) {
-        case VpRequestQueryType.didAuth:
-          return this.verifyVpRequestTypeDidAuth(presentation);
-        case VpRequestQueryType.presentationDefinition:
-          return this.verifyVpRequestTypePresentationDefinition(presentation, vpQuery.credentialQuery);
-        default:
-          return ['Unknown request query type'];
-      }
-    });
-
-    return [...partialErrors, ...commonErrors];
-  }
-
   /**
    * Process a presentation submission.
-   * Check the correctness of the presentation against the VP Request Credential Queries.
-   * Does NOT check signatures.
    * @param presentation
+   * @param verifier
    */
-  public processPresentation(presentation: VerifiablePresentation): {
-    response: ExchangeResponseDto;
-    callback: CallbackConfiguration[];
-  } {
-    const errors = this.validatePresentation(presentation);
+  public async processPresentation(
+    presentation: VerifiablePresentation,
+    verifier: VpRequestSubmissionVerifier
+  ): Promise<{ response: ExchangeResponseDto; callback: CallbackConfiguration[] }> {
+    const verificationResult = await verifier.verifyVpRequestSubmission(presentation, this.vpRequest);
+    const errors = verificationResult.errors;
 
     if (errors.length > 0) {
       return {
